Remove base execution, use serial_execution instead
